Bugfix: handle a request after it is ended.
I noticed an issue when trying ng-apimock with angular-cli (as described in ANGULAR_CLI.md):
When a POST request with a body is done then the mocked response is returned before the request has been ended.
This often causes issues in the Angular proxy, either by throwing an error "[HPM] Error occurred while trying to proxy request" or
by responding with half of the mocked JSON response.

Waiting on the 'end' event of the HTTP-request fixed this issue.